Fix "wrong issue in hero" bug
When a new issue is released there is a 30 - 45 min delay introduced by our caching the adobe XML feed on our own server.

During that delay, the "latest issue" is filtered out from the libraryService.get_visible() due to having no adobe data, causing the hero to display the *next* issue in the sort order.

This fix allows the previous Issue that was in the Hero to stay until the adobe feed is updated.